+++ layers/API/packages/api/src/ObjectModels/SchemaDefinition/InputObjectTypeSchemaDefinitionProvider.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace PoPAPI\API\ObjectModels\SchemaDefinition;
 
-use PoP\ComponentModel\TypeResolvers\InputObjectType\InputObjectTypeResolverInterface;
-use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
 use PoPAPI\API\Schema\SchemaDefinition;
 use PoPAPI\API\Schema\SchemaDefinitionHelpers;
 use PoPAPI\API\Schema\TypeKinds;
+use PoP\ComponentModel\TypeResolvers\InputObjectType\InputObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\InputObjectType\OneofInputObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
 
 class InputObjectTypeSchemaDefinitionProvider extends AbstractNamedTypeSchemaDefinitionProvider
 {
@@ -31,6 +32,7 @@ class InputObjectTypeSchemaDefinitionProvider extends AbstractNamedTypeSchemaDef
         $schemaDefinition = parent::getSchemaDefinition();
 
         $this->addInputFieldSchemaDefinitions($schemaDefinition);
+        $this->addIsOneOfSchemaDefinitions($schemaDefinition);
 
         return $schemaDefinition;
     }
@@ -59,5 +61,15 @@ class InputObjectTypeSchemaDefinitionProvider extends AbstractNamedTypeSchemaDef
 
             $schemaDefinition[SchemaDefinition::INPUT_FIELDS][$inputFieldName] = $inputFieldSchemaDefinition;
         }
+    }
+
+    /**
+     * "oneOf" Input Objects
+     *
+     * @param array<string,mixed> $schemaDefinition
+     */
+    final protected function addIsOneOfSchemaDefinitions(array &$schemaDefinition): void
+    {
+        $schemaDefinition[SchemaDefinition::IS_ONE_OF] = $this->inputObjectTypeResolver instanceof OneofInputObjectTypeResolverInterface;
     }
 }