+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -36,6 +36,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Validate the domain in the active license corresponds to the current site (for commercial extensions) (#3192)
 - Don't trigger license check if never run before (#3194)
 - Set `WithMeta.metaKeys` as sensitive field (#3197)
+- Move `isOneOf` introspection field out of `extensions` (as feature has been merged to GraphQL spec) (#3198)
 
 ### Fixed